Initial Commit (Read Extended Description)
This will not build currently. However a release DLL will accompany it so that people may still utilize it if that is the goal to begin with

Whats Missing:
- Dear ImGui Library
- Minhook Library

I plan to get to this today at some point , but I figured I would go ahead and get the ball rolling.